// src/models/RouterNamesModel.ts
export enum ROUTES {
  APP_NAVIGATION = "APP_NAVIGATION",
  GUEST_NAVIGATION = "GUEST_NAVIGATION",
}

export enum EAppScreenList {
  HEALTH_STACK_SCREEN = "HEALTH_STACK_SCREEN",
  CALENDAR_STACK_SCREEN = "CALENDAR_STACK_SCREEN",
  CHAT_STACK_SCREEN = "CHAT_STACK_SCREEN",
  MORE_STACK_SCREEN = "MORE_STACK_SCREEN",
}

export enum EGuestScreenList {
  LOGIN_SCREEN = "LOGIN_SCREEN",
  SUCCESS_SCREEN = "SUCCESS_SCREEN",
  VERIFICATION_SCREEN = "VERIFICATION_SCREEN",
  SELECT_CLINIC_SCREEN = "SELECT_CLINIC_SCREEN",
  FORGOT_PASSWORD_SCREEN = "FORGOT_PASSWORD_SCREEN",
  ACTIVATION_CODE_SCREEN = "ACTIVATION_CODE_SCREEN",
  PERSONAL_DETAILS_SCREEN = "PERSONAL_DETAILS_SCREEN",
  TERMS_CONDITIONS_SCREEN = "TERMS_CONDITIONS_SCREEN",
  PERSONAL_ADDRESS_SCREEN = "PERSONAL_ADDRESS_SCREEN",
  ACTIVATION_QUESTION_SCREEN = "ACTIVATION_QUESTION_SCREEN",
  CREATE_NEW_PASSWORD_SCREEN = "CREATE_NEW_PASSWORD_SCREEN",
}

export enum EHealthScreenList {
  USER_PROFILE_SCREEN = "USER_PROFILE_SCREEN",
  UPDATE_PROFILE_SCREEN = "UPDATE_PROFILE_SCREEN",
  HEALTH_STATS_DETAIL_SCREEN = "HEALTH_STATS_DETAIL_SCREEN",
  HEALTH_STATS_OVERVIEW_SCREEN = "HEALTH_STATS_OVERVIEW_SCREEN",
  HEALTH_SINGLE_STAT_DETAIL_SCREEN = "HEALTH_SINGLE_STAT_DETAIL_SCREEN",
}

export enum EAppointmentScreenList {
  BOOK_APPOINTMENT_SCREEN = "BOOK_APPOINTMENT_SCREEN",
  APPOINTMENT_DETAIL_SCREEN = "APPOINTMENT_DETAIL_SCREEN",
  APPOINTMENT_ACTIONS_SCREEN = "APPOINTMENT_ACTIONS_SCREEN",
  STEP_ONE_APPOINTMENT_SCREEN = "STEP_ONE_APPOINTMENT_SCREEN",
  STEP_TWO_APPOINTMENT_SCREEN = "STEP_TWO_APPOINTMENT_SCREEN",
  PRACTITIONER_PROFILE_SCREEN = "PRACTITIONER_PROFILE_SCREEN",
  STEP_THREE_APPOINTMENT_SCREEN = "STEP_THREE_APPOINTMENT_SCREEN",
  APPOINTMENT_VIDEO_CALL_SCREEN = "APPOINTMENT_VIDEO_CALL_SCREEN",
  VERIFY_APPOINTMENT_BOOK_SCREEN = "VERIFY_APPOINTMENT_BOOK_SCREEN",
  APPOINTMENT_BOOK_SUCCESS_SCREEN = "APPOINTMENT_BOOK_SUCCESS_SCREEN",
  UPCOMING_PREVIOUS_APPOINTMENT_SCREEN = "UPCOMING_PREVIOUS_APPOINTMENT_SCREEN",
}

export enum EChatScreenList {
  CHAT_SCREEN = "CHAT_SCREEN",
  CHAT_ROOM_SCREEN = "CHAT_ROOM_SCREEN",
  VIDEO_CALL_SCREEN = "VIDEO_CALL_SCREEN",
}

export enum EMoreScreenList {
  MORE_SCREEN = "MORE_SCREEN",
  SETTING_SCREEN = "SETTING_SCREEN",
  PATIENT_PROFILE_SCREEN = "PATIENT_PROFILE_SCREEN",
}